+++ src-tauri/src/models.rs
@@ -1,0 +1,7 @@
+use serde::{Deserialize, Serialize};
+
+#[derive(Clone, Debug, Serialize, Deserialize)]
+pub struct SanitizedSegment {
+    pub name: String,
+    pub fsname: String,
+}
