+++ src-tauri/src/space/mod.rs
@@ -30,7 +30,7 @@ pub fn create_space(dto: CreateSpaceDto, sharedstate: &mut SharedState) -> Resul
         )));
     }
 
-    let space_dirname = utils::sanitize_pathseg(&dto.name);
+    let space_dirname = utils::sanitize_name(&dto.name);
     let space_abspath = location.join(&space_dirname);
     let mut spacerefs = store::get_spacerefs();
